# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sphere English LMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /courses:
    get:
      operationId: getCourses
      tags: [courses]
      summary: List all courses
      parameters:
        - name: level
          in: query
          schema:
            type: ["string", "null"]
            enum: [A1, A2, B1, B2, C1, C2, null]
        - name: teacherId
          in: query
          schema:
            type: ["integer", "null"]
        - name: isActive
          in: query
          schema:
            type: ["boolean", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"

    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a new course (admin/teacher)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseRequest"
      responses:
        "201":
          description: Course created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get course by ID with modules and lessons
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"

    patch:
      operationId: updateCourse
      tags: [courses]
      summary: Update course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCourseRequest"
      responses:
        "200":
          description: Updated course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"

    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /courses/{id}/enroll:
    post:
      operationId: enrollCourse
      tags: [courses]
      summary: Enroll student in a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollRequest"
      responses:
        "200":
          description: Enrollment successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /courses/{id}/unenroll:
    post:
      operationId: unenrollCourse
      tags: [courses]
      summary: Unenroll student from a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollRequest"
      responses:
        "200":
          description: Unenrollment successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /certificates/{id}:
    get:
      operationId: getCertificate
      tags: [certificates]
      summary: Get certificate by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Certificate details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certificate"
  /certificates/verify/{code}:
    get:
      operationId: verifyCertificate
      tags: [certificates]
      summary: Verify a certificate by its code
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Certificate verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CertificateVerification"
  /certificates/issue:
    post:
      operationId: issueCertificate
      tags: [certificates]
      summary: Issue a certificate for a completed course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueCertificateRequest"
      responses:
        "201":
          description: Certificate issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certificate"

  # Messages
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    # Auth
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [admin, teacher, student]
        phone:
          type: ["string", "null"]
      required:
        - email
        - password
        - firstName
        - lastName
        - role
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"
        token:
          type: string
      required:
        - user
        - token

    # Users
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        role:
          type: string
          enum: [admin, teacher, student]
        firstName:
          type: string
        lastName:
          type: string
        avatar:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        createdAt:
          type: string
        currentLevel:
          type: ["string", "null"]
          enum: [A1, A2, B1, B2, C1, C2, null]
        totalPoints:
          type: ["integer", "null"]
        streak:
          type: ["integer", "null"]
        badges:
          type: ["array", "null"]
          items:
            type: string
      required:
        - id
        - email
        - role
        - firstName
        - lastName
        - createdAt
...
    Course:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        level:
          type: string
          enum: [A1, A2, B1, B2, C1, C2]
        teacherId:
          type: integer
        teacherName:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        isActive:
          type: boolean
        enrolledCount:
          type: integer
        totalLessons:
          type: integer
        createdAt:
          type: string
        imageUrl:
          type: ["string", "null"]
      required:
        - id
        - title
        - description
        - level
        - teacherId
        - isActive
        - enrolledCount
        - totalLessons
        - createdAt
    CourseDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        level:
          type: string
          enum: [A1, A2, B1, B2, C1, C2]
        teacherId:
          type: integer
        teacherName:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        isActive:
          type: boolean
        enrolledCount:
          type: integer
        totalLessons:
          type: integer
        createdAt:
          type: string
        imageUrl:
          type: ["string", "null"]
        modules:
          type: array
          items:
            $ref: "#/components/schemas/ModuleWithLessons"
        isEnrolled:
          type: boolean
        completionPercentage:
          type: ["number", "null"]
      required:
        - id
        - title
        - description
        - level
        - teacherId
        - isActive
        - enrolledCount
        - totalLessons
        - createdAt
        - modules
        - isEnrolled
    CreateCourseRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        level:
          type: string
          enum: [A1, A2, B1, B2, C1, C2]
        teacherId:
          type: ["integer", "null"]
        price:
          type: ["number", "null"]
        isActive:
          type: boolean
        imageUrl:
          type: ["string", "null"]
      required:
        - title
        - description
        - level
        - isActive
    UpdateCourseRequest:
      type: object
      properties:
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        level:
          type: ["string", "null"]
          enum: [A1, A2, B1, B2, C1, C2, null]
        teacherId:
          type: ["integer", "null"]
        price:
          type: ["number", "null"]
        isActive:
          type: ["boolean", "null"]
        imageUrl:
          type: ["string", "null"]
    EnrollRequest:
      type: object
      properties:
        studentId:
          type: ["integer", "null"]
      # studentId is optional; defaults to current user if not provided

    # Modules
...
    ModuleWithLessons:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        order:
          type: integer
        lessons:
          type: array
          items:
            $ref: "#/components/schemas/Lesson"
      required:
        - id
        - courseId
        - title
        - order
        - lessons
...
    Lesson:
      type: object
      properties:
        id:
          type: integer
        moduleId:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [video, document, quiz, live, audio, text, image]
        content:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        order:
          type: integer
        isCompleted:
          type: ["boolean", "null"]
      required:
        - id
        - moduleId
        - title
        - type
        - order
...
    Certificate:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: ["string", "null"]
        courseId:
          type: integer
        courseTitle:
          type: string
        level:
          type: string
        issuedAt:
          type: string
        verificationCode:
          type: string
        teacherName:
          type: ["string", "null"]
      required:
        - id
        - userId
        - courseId
        - courseTitle
        - level
        - issuedAt
        - verificationCode
    CertificateVerification:
      type: object
      properties:
        valid:
          type: boolean
        certificate:
          $ref: "#/components/schemas/Certificate"
      required:
        - valid
    IssueCertificateRequest:
      type: object
      properties:
        userId:
          type: integer
        courseId:
          type: integer
      required:
        - userId
        - courseId

    # Messages
...